// frontend/src/services/model/userResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Football Pool API
 * OpenAPI spec version: 1.0.0
 */
import type { PlayerResponse } from "./playerResponse";

export interface UserResponse {
  id: number;
  name: string;
  email: string;
  role: string;
  player?: PlayerResponse;
  created_at: string;
  updated_at: string;
}